Fix bug in exporting dependencies of initial tables
Set all initial tables and subject conditions to be day 0.
The bug this fixes is where we had a list of initial tables:

foos
bars
blarghs

and then the subject table:

samples

So foos would be day 1. But by the time we would start exporting the dependencies of foos,
we would already be at day 5. Then the query was only looking for things in the previous day (day 4),
missing the fact that foos was exported on day 1, not day 4.

This sets initial tables and the subject table to all be exported on day 0 together.